+++ app/src/main/kotlin/br/com/youse/forms/samples/login/livedata/BindingAdapters.kt
@@ -21,7 +21,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
-package br.com.youse.forms.samples.livedata
+package br.com.youse.forms.samples.login.livedata
 
 import android.databinding.BindingAdapter
 import android.view.View
